[FIX] Post-merge test fixes, run locally, need to check remote
