+++ autocompliance/setup.sh
