+++ autocompliance/src/demo_functions.py
@@ -2,7 +2,9 @@
 
 # Author: @andrewk10
 
-# Importing file for file based functionality.
+# Importing argparse for command-line option parsing
+import argparse
+# Importing file for file based functionality
 import autocompliance.src.file as file
 # Importing logging to safely log sensitive, error or debug info.
 import logging
@@ -83,6 +85,57 @@ class DemoFunctions:
             return None
 
 
+def parse_arguments(arguments):
+    """
+    This function will parse arguments and is even used to dummy arguments for
+    tests when needed
+    :param arguments: The arguments to be parsed
+    """
+
+    # Argument parser for handling arguments.
+    parser = argparse.ArgumentParser(description=strings.DESCRIPTION)
+    # Adding the target  file option to the parser.
+    parser.add_argument(
+        strings.IP_FILE_OPT_SHORT, strings.IP_FILE_OPT_LONG,
+        dest='target', help=strings.IP_FILE_HELP, type=str)
+    # Adding the username option to the parser.
+    parser.add_argument(
+        strings.USERNAME_OPT_SHORT, strings.USERNAME_OPT_LONG,
+        dest='username', help=strings.USERNAME_HELP, type=str)
+    # Adding the password file option to the parser.
+    parser.add_argument(
+        strings.PW_FILE_OPT_SHORT, strings.PW_FILE_OPT_LONG,
+        dest="pw_file", help=strings.PW_FILE_HELP, type=str)
+    # Adding the port option to the parser.
+    parser.add_argument(
+        strings.PORT_OPT_SHORT, strings.PORT_OPT_LONG,
+        dest='ports', help=strings.PORT_HELP, type=str)
+    # Adding the lan option to the parser.
+    parser.add_argument(
+        strings.LAN_OPT_SHORT, strings.LAN_OPT_LONG, action='store_true',
+        help=strings.LAN_HELP)
+    # Adding the propagate option to the parser.
+    parser.add_argument(
+        strings.PROP_OPT_SHORT, strings.PROP_OPT_LONG, action='store_true',
+        help=strings.PROP_HELP)
+    # Adding the transfer file option to the parser.
+    parser.add_argument(
+        strings.PROP_FILE_OPT_SHORT, strings.PROP_FILE_OPT_LONG,
+        dest='propagate_file', help=strings.PROP_FILE_HELP, type=str)
+
+    # Parsing the arguments.
+    arguments = parser.parse_args(arguments)
+    return arguments
+
+
+def exit_and_show_instructions():
+    """
+    This function will print the help screen and show an exit prompt.
+    """
+    print(strings_functions.help_output())
+    print(strings.EXITING)
+
+
 def remove_duplicates_in_list(list_with_duplicates):
     """
     This function simply removes duplicate entries in a given list
@@ -94,11 +147,3 @@ def remove_duplicates_in_list(list_with_duplicates):
         if ip not in deduplicated_list:
             deduplicated_list.append(ip)
     return deduplicated_list
-
-
-def exit_and_show_instructions():
-    """
-    This function will print the help screen and show an exit prompt.
-    """
-    print(strings_functions.help_output())
-    print(strings.EXITING)
