+++ autocompliance/tests.sh
